requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fastapi==0.92.0
fonttools==4.38.0
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
ipython==8.11.0
jedi==0.18.2
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.5
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.25.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
zope.event==4.6
zope.interface==5.5.2
